navigates to Stores from Homes moves Stores to own component file imports Stores into App
