+++ src/App.js
@@ -6,6 +6,8 @@ import {
 } from 'react-router-dom';
 
 import Home from './components/Home';
+import { Stores } from './components/Stores';
+
 import './App.scss';
 
 const App = () => (
@@ -22,40 +24,6 @@ const App = () => (
       <Route path='/stores' component={Stores}/>
     </nav>
   </Router>
-);
-
-const Stores = ({ match }) => (
-  <div>
-    <h2>Stores</h2>
-    <ul>
-      <li>
-        <Link to={`${match.url}/rendering`}>
-          Rendering with React
-        </Link>
-      </li>
-      <li>
-        <Link to={`${match.url}/components`}>
-          Components
-        </Link>
-      </li>
-      <li>
-        <Link to={`${match.url}/props-v-state`}>
-          Props v. State
-        </Link>
-      </li>
-    </ul>
-
-    <Route path={`${match.url}/:topicId`} component={Topic}/>
-    <Route exact path={match.url} render={() => (
-      <h3>Please select a topic.</h3>
-    )}/>
-  </div>
-);
-
-const Topic = ({ match }) => (
-  <div>
-    <h3>{match.params.topicId}</h3>
-  </div>
 );
 
 export default App;
